pulled duckdb connection management to its own helper
